feat(#S11P12D109-349) : 리뷰 상세

See merge request s11-webmobile4-sub2/S11P12D109!127

+++ Colorpl-Android/presentation/src/main/java/com/presentation/component/adapter/feed/FeedAdapter.kt
@@ -9,7 +9,7 @@ import com.domain.model.Feed
 import com.presentation.base.BaseDiffUtil
 
 class FeedAdapter(
-    private val onFeedContentClickListener: () -> Unit,
+    private val onFeedContentClickListener: (Int) -> Unit,
     private val onCommentClickListener: () -> Unit,
     private val onEmotionClickListener: () -> Unit,
     private val onReportClickListener: () -> Unit,
